+++ src/DaysView.js
@@ -2,11 +2,10 @@
 
 var React = require('react'),
 	createClass = require('create-react-class'),
-	moment = require('moment'),
-	onClickOutside = require('react-onclickoutside').default
-	;
+	moment = require('moment')
+;
 
-var DateTimePickerDays = onClickOutside( createClass({
+var DateTimePickerDays = createClass({
 	render: function() {
 		var footer = this.renderFooter(),
 			date = this.props.viewDate,
@@ -134,11 +133,7 @@ var DateTimePickerDays = onClickOutside( createClass({
 
 	alwaysValidDate: function() {
 		return 1;
-	},
-
-	handleClickOutside: function() {
-		this.props.handleClickOutside();
 	}
-}));
+});
 
 module.exports = DateTimePickerDays;