+++ src/YearsView.js
@@ -1,11 +1,10 @@
 'use strict';
 
 var React = require('react'),
-	createClass = require('create-react-class'),
-	onClickOutside = require('react-onclickoutside').default
+	createClass = require('create-react-class')
 	;
 
-var DateTimePickerYears = onClickOutside( createClass({
+var DateTimePickerYears = createClass({
 	render: function() {
 		var year = parseInt( this.props.viewDate.year() / 10, 10 ) * 10;
 
@@ -96,10 +95,6 @@ var DateTimePickerYears = onClickOutside( createClass({
 	alwaysValidDate: function() {
 		return 1;
 	},
-
-	handleClickOutside: function() {
-		this.props.handleClickOutside();
-	}
-}));
+});
 
 module.exports = DateTimePickerYears;