+++ webpack.config.js
@@ -23,7 +23,8 @@ module.exports = {
   externals: {
     'react': 'React',
     'react-dom': 'ReactDOM',
-    'moment': 'moment'
+    'moment': 'moment',
+    'moment-timezone': 'moment-timezone'
   },
 
   plugins: plugins